feat: MET-2055 update governance vote
